sysutils/edk2: make bhyve's bootorder feature work
Bhyve supports a bootindex property for bootable devices. When adding it
to at least one device, bhyve creates a bootorder and passes it over the
Qemu FwCfg interface to the guest. The guest firmware has to pick up the
FwCfg file and has to apply the bootorder. Otherwise, the bootindex
property won't have any effect. Recently, a commit was merged by EDKII
to support that feature and we can simply cherry-pick it.

PR:			279720
Reviewed by:		<If someone else reviewed your modification.>
Approved by:		<If you needed approval for this commit.>
Sponsored by:		Beckhoff Automation GmbH & Co. KG
Differential Revision:	https://reviews.freebsd.org/D45768